requirements.txt
# Core Pipeline
dlt[duckdb]>=0.4.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0

# Orchestration
apache-airflow>=2.8.0
astronomer-cosmos>=1.3.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Data Sources
requests>=2.31.0
praw>=7.7.0
finnhub-python>=2.4.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
